Rithekha-3122225001106-otx/requirements.txt
requests==2.31.0
pymongo==4.6.0
python-dotenv==1.0.0